brings findbugs count from 266 -> 250 bugs by fixing some scopes, removing some dead code, and not implementing Serializable unnecessarily.  patch by Michael Greene; reviewed for CASSANDRA-338 by jbellis
git-svn-id: https://svn.apache.org/repos/asf/incubator/cassandra/trunk@800547 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/cassandra/gms/Gossiper.java
@@ -457,7 +457,7 @@ public class Gossiper implements IFailureDetectionEventListener, IEndPointStateC
             else
             {
                 /* Gossip with the seed with some probability. */
-                double probability = seeds_.size() / ( liveEndpoints_.size() + unreachableEndpoints_.size() );
+                double probability = seeds_.size() / (double)( liveEndpoints_.size() + unreachableEndpoints_.size() );
                 double randDbl = random_.nextDouble();
                 if ( randDbl <= probability )
                     sendGossip(message, seeds_);
